requirements.txt
Langchain
pypdf
chromadb
openai
python-dotenv